Controlers details, postman test

+++ src/main/java/co/edu/uniandes/dse/treeking/dto/LocationDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class LocationDTO {
-    private float latitude;
-    private float longitude;
+    private Double latitude;
+    private Double longitude;
 }
